backend/requirements.txt
asgiref==3.11.1
Django==5.2.11
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.16.1
python-decouple==3.8
python-dotenv==1.2.1
sqlparse==0.5.5
typing_extensions==4.15.0
